% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splot-arrows.R
\name{arrow_base_midpoint}
\alias{arrow_base_midpoint}
\title{Calculate Arrow Base Midpoint}
\usage{
arrow_base_midpoint(x, y, angle, size, width = 0.5)
}
\arguments{
\item{x}{Arrow tip x coordinate.}

\item{y}{Arrow tip y coordinate.}

\item{angle}{Angle of incoming edge (radians).}

\item{size}{Arrow size in user coordinates.}

\item{width}{Arrow width ratio (default 0.5).}
}
\value{
List with x, y coordinates of the arrow base midpoint.
}
\description{
Returns the midpoint between the arrow wings (where the curve should end).
This is used to connect the edge line to the back of the arrow head.
}
\keyword{internal}
