% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\name{community_sizes}
\alias{community_sizes}
\title{Get Community Sizes}
\usage{
community_sizes(x)
}
\arguments{
\item{x}{A cograph_communities object}
}
\value{
Named integer vector of community sizes
}
\description{
Get Community Sizes
}
