% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-utils.R
\name{.check_validation_available}
\alias{.check_validation_available}
\title{Check if Validation Method is Available}
\usage{
.check_validation_available(x, method)
}
\arguments{
\item{x}{Network object.}

\item{method}{Validation method name.}
}
\value{
Logical.
}
\description{
Check if Validation Method is Available
}
\keyword{internal}
