% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-metrics.R
\name{.cluster_summary_to_tna}
\alias{.cluster_summary_to_tna}
\title{Convert cluster_summary to tna object}
\usage{
.cluster_summary_to_tna(x)
}
\description{
Convert cluster_summary to tna object
}
\keyword{internal}
