% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-utils.R
\name{.tna_estimator}
\alias{.tna_estimator}
\title{TNA Weight Estimator}
\usage{
.tna_estimator(data, type = "relative", scaling = "none")
}
\arguments{
\item{data}{TNA sequence data (tna_seq_data class).}

\item{type}{Weight type ("relative", "frequency", etc.).}

\item{scaling}{Scaling method.}
}
\value{
Weight matrix.
}
\description{
Internal estimator function for TNA sequence data.
Matches TNA's internal weight computation exactly.
}
\keyword{internal}
