% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from-qgraph.R
\name{map_qgraph_shape}
\alias{map_qgraph_shape}
\title{Map qgraph shape names to Sonnet equivalents}
\usage{
map_qgraph_shape(shapes)
}
\arguments{
\item{shapes}{Character vector of qgraph shape names}
}
\value{
Character vector of Sonnet shape names
}
\description{
Map qgraph shape names to Sonnet equivalents
}
\keyword{internal}
