% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\name{membership.cograph_communities}
\alias{membership.cograph_communities}
\title{Get Community Membership}
\usage{
membership.cograph_communities(x)
}
\arguments{
\item{x}{A cograph_communities object}
}
\value{
Named integer vector of community assignments
}
\description{
Get Community Membership
}
