% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\name{modularity.cograph_communities}
\alias{modularity.cograph_communities}
\title{Get Modularity Score}
\usage{
modularity.cograph_communities(x, ...)
}
\arguments{
\item{x}{A cograph_communities object}
}
\value{
Numeric modularity value
}
\description{
Get Modularity Score
}
