% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\name{n_communities}
\alias{n_communities}
\title{Get Number of Communities}
\usage{
n_communities(x)
}
\arguments{
\item{x}{A cograph_communities object}
}
\value{
Integer count of communities
}
\description{
Get Number of Communities
}
