% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{permutation_test}
\alias{permutation_test}
\title{Permutation Test}
\usage{
permutation_test(
  x,
  y,
  adjust = "none",
  iter = 1000,
  paired = FALSE,
  level = 0.05,
  measures = character(0),
  ...
)
}
\arguments{
\item{x}{First tna object.}

\item{y}{Second tna object.}

\item{adjust}{P-value adjustment method. Default "none".}

\item{iter}{Number of iterations. Default 1000.}

\item{paired}{Logical. Paired test? Default FALSE.}

\item{level}{Significance level. Default 0.05.}

\item{measures}{Centrality measures to compare.}

\item{...}{Additional arguments.}
}
\value{
A tna_permutation object.
}
\description{
Permutation Test
}
